requirements.txt
coremltools==7.0
matplotlib==3.8.0
numpy==1.26.1
onnxruntime==1.16.1
opencv_python_headless==4.8.1.78
pandas==2.1.2
Pillow==10.1.0
py_cpuinfo==9.0.0
pytest==7.4.3
PyYAML==6.0.1
scikit_learn==1.3.2
tensorflow==2.14.0
tensorflow_intel==2.14.0
tqdm==4.66.1
